refactor(tidy): replace runtime bootstrap with bootstrap owners
