+++ backend/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Shared code for FastAPI app shells."""
