+++ backend/monitor/api/http/dependencies.py
@@ -2,7 +2,12 @@
 
 from __future__ import annotations
 
+from fastapi import FastAPI, Request
+
 from backend.identity.auth.user_resolution import get_current_user_id
-from backend.runtime_bootstrap.request_app import get_app
 
 __all__ = ["get_app", "get_current_user_id"]
+
+
+async def get_app(request: Request) -> FastAPI:
+    return request.app
