--- backend/runtime_bootstrap/request_app.py
@@ -1,7 +1,0 @@
-"""Neutral request->app dependency helper."""
-
-from fastapi import FastAPI, Request
-
-
-async def get_app(request: Request) -> FastAPI:
-    return request.app
