+++ backend/threads/file_channel.py
@@ -6,8 +6,8 @@ import logging
 from dataclasses import dataclass
 from pathlib import Path
 
-from backend.runtime_bootstrap.storage_container_cache import get_storage_container as _get_container
 from config.user_paths import user_home_path
+from storage.container_cache import get_storage_container as _get_container
 
 logger = logging.getLogger(__name__)
